JsAsesoriaG/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
cryptography==45.0.5
Django==5.2.4
django-adminlte-3==0.1.6
django-allauth==65.10.0
django-crispy-forms==2.4
django-environ==0.12.0
django-jazzmin==3.0.1
idna==3.10
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
